Add circle-square and circle-circle collision detection

+++ src/GameEngine/Collision/CircleCC.h
@@ -5,12 +5,13 @@
 #include <cstdint>
 #include "CollisionComponent.h"
 
+// TODO: add position offset
 class CircleCC : public CollisionComponent{
 public:
-	CircleCC(uint16_t radius);
-	uint16_t getRadius() const;
+	CircleCC(float radius);
+	float getRadius() const;
 private:
-	uint16_t radius;
+	float radius;
 };
